+++ CHANGELOG.md
@@ -2,9 +2,12 @@
 
 ## Version 0.1.1
 
-* Enhancements to GandiV5::Email::Mailbox:
+* Enhancements to GandiV5::Email::Mailbox
   * .create now returns created mailbox
   * .create now checks for available slots and a valid type has been passed
+* Enhancements to GandiV5::Email::Mailbox::Responder
+  * Add #enable(message:, ends_at:, starts_at: Time.now) to enable the auto responder in Gandi
+  * Add #disable to disable the auto responder in Gandi
 * Add support for ruby 2.6.3
 
 ## Version 0.1.0